+++ TJC.Logging/Providers/ConsoleLoggerProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace TJC.Logging.Providers;
+
+internal class ConsoleLoggerProvider : ILoggerProvider
+{
+    public ILogger CreateLogger(string categoryName) =>
+        new ConsoleLogger(categoryName);
+
+    public void Dispose() { }
+}